Move SpecimenApiController to specimen-main
