+++ study/api-src/org/labkey/api/specimen/actions/SpecimenApiForm.java
@@ -1,0 +1,21 @@
+package org.labkey.api.specimen.actions;
+
+import org.labkey.api.action.HasViewContext;
+import org.labkey.api.view.ViewContext;
+
+public class SpecimenApiForm implements HasViewContext
+{
+    private ViewContext _viewContext;
+
+    @Override
+    public ViewContext getViewContext()
+    {
+        return _viewContext;
+    }
+
+    @Override
+    public void setViewContext(ViewContext viewContext)
+    {
+        _viewContext = viewContext;
+    }
+}
